requirements.txt
numpy
pytorch
torchvision
scikit-learn
matplotlib
pillow
sklearn
torchvision
torch
